Fix Uncaught Mutex Interrupt Exceptions
The TimeoutEnforcer used ruby's Mutex class which is interruptible
and throws a special exception. This switches the code to use
ReentrantLock (which JRuby's Mutex is based on) directly, and use
ReentrantLock.lock instead of ReentrantLock.lockInterruptibly

This also adds a small note about the placement of thread.interrupted.

Fixes https://github.com/logstash-plugins/logstash-filter-grok/issues/97

Fixes #98